# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Старт API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectDifficulty"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProjectFormat"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          },
          {
            "name": "skill_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Skill Id"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "difficulty"
              ],
              "type": "string",
              "default": "newest",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectListItem"
                  },
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/recommended": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Recommended Projects",
        "operationId": "list_recommended_projects_projects_recommended_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectRecommendation"
                  },
                  "type": "array",
                  "title": "Response List Recommended Projects Projects Recommended Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Get My Teams",
        "operationId": "get_my_teams_me_teams_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TeamRead"
                  },
                  "type": "array",
                  "title": "Response Get My Teams Me Teams Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/team": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Get My Team",
        "operationId": "get_my_team_me_team_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/me/projects/{project_id}/submission": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get My Submission",
        "operationId": "get_my_submission_me_projects__project_id__submission_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "submissions"
        ],
        "summary": "Submit Project Result",
        "operationId": "submit_project_result_me_projects__project_id__submission_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "type",
          "verified"
        ],
        "title": "OrganizationRead"
      },
...
      "ProjectDifficulty": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced"
        ],
        "title": "ProjectDifficulty"
      },
...
      "ProjectFormat": {
        "type": "string",
        "enum": [
          "online",
          "offline",
          "hybrid"
        ],
        "title": "ProjectFormat"
      },
      "ProjectListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProjectDifficulty"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "format": {
            "$ref": "#/components/schemas/ProjectFormat"
          },
          "deadline": {
            "type": "string",
            "title": "Deadline"
          },
          "participant_limit": {
            "type": "integer",
            "title": "Participant Limit"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty",
          "status",
          "format",
          "deadline",
          "participant_limit",
          "organization"
        ],
        "title": "ProjectListItem"
      },
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProjectDifficulty"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "format": {
            "$ref": "#/components/schemas/ProjectFormat"
          },
          "deadline": {
            "type": "string",
            "title": "Deadline"
          },
          "participant_limit": {
            "type": "integer",
            "title": "Participant Limit"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          },
          "expected_result": {
            "type": "string",
            "title": "Expected Result"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/ProjectRoleRead"
            },
            "type": "array",
            "title": "Roles"
          },
          "required_skills": {
            "items": {
              "$ref": "#/components/schemas/ProjectSkillRead"
            },
            "type": "array",
            "title": "Required Skills"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty",
          "status",
          "format",
          "deadline",
          "participant_limit",
          "organization",
          "expected_result",
          "roles",
          "required_skills"
        ],
        "title": "ProjectRead"
      },
      "ProjectRecommendation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProjectDifficulty"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "format": {
            "$ref": "#/components/schemas/ProjectFormat"
          },
          "deadline": {
            "type": "string",
            "title": "Deadline"
          },
          "participant_limit": {
            "type": "integer",
            "title": "Participant Limit"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationRead"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Breakdown"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "difficulty",
          "status",
          "format",
          "deadline",
          "participant_limit",
          "organization"
        ],
        "title": "ProjectRecommendation"
      },
...
      "ProjectRoleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "slots": {
            "type": "integer",
            "title": "Slots"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "slots"
        ],
        "title": "ProjectRoleRead"
      },
...
      "ProjectSkillRead": {
        "properties": {
          "skill": {
            "$ref": "#/components/schemas/SkillRead"
          },
          "required_level": {
            "$ref": "#/components/schemas/SkillLevel"
          }
        },
        "type": "object",
        "required": [
          "skill",
          "required_level"
        ],
        "title": "ProjectSkillRead"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "draft",
          "open",
          "in_progress",
          "completed"
        ],
        "title": "ProjectStatus"
      },
...
      "ProjectSubmissionStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "revision_requested",
          "approved",
          "rejected"
        ],
        "title": "ProjectSubmissionStatus"
      },
      "SkillLevel": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced"
        ],
        "title": "SkillLevel"
      },
...
      "SkillRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category"
        ],
        "title": "SkillRead"
      },
      "SubmissionCreate": {
        "properties": {
          "summary": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 10,
            "title": "Summary"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "SubmissionCreate"
      },
      "SubmissionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectSubmissionStatus"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "user_id",
          "summary",
          "result_url",
          "status",
          "review_note",
          "submitted_at",
          "reviewed_at"
        ],
        "title": "SubmissionRead"
      },
...
      "TeamMemberRead": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/TeamMemberUserRead"
          },
          "role_title": {
            "type": "string",
            "title": "Role Title"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "user",
          "role_title",
          "joined_at"
        ],
        "title": "TeamMemberRead"
      },
      "TeamMemberUserRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar_url"
        ],
        "title": "TeamMemberUserRead"
      },
      "TeamRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectListItem"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/TeamMemberRead"
            },
            "type": "array",
            "title": "Members"
          },
          "team_chat_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Chat Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "project",
          "members"
        ],
        "title": "TeamRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}